pyproject.toml
[project]
name = "cloudstorage"
version = "0.1.0"
description = ""
authors = [
    {name = "maxdikun",email = "a10levsion@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt[cryptography] (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
